Refactor mock tool test interactor for reuse.
